Rename backends: web -> webgpu, direct -> wgpu_core
