+++ wgpu/src/backend/mod.rs
@@ -1,9 +1,9 @@
 #[cfg(webgpu)]
-mod web;
+mod webgpu;
 #[cfg(webgpu)]
-pub(crate) use web::Context;
+pub(crate) use webgpu::Context;
 
 #[cfg(not(webgpu))]
-mod direct;
+mod wgpu_core;
 #[cfg(not(webgpu))]
-pub(crate) use direct::Context;
+pub(crate) use wgpu_core::Context;
